Combined Division with External Company. Still is 2 tabs though for adding an industry.
I added the class files for ExternalCompany, Correspondence, Industry and Contact

All of them are erroring out though due to this:
Columns of type 'INTEGER' cannot hold values of type 'CHAR'.
I cast the ints to strings like you have it, but it does not like that in the database. Btw I still cannot see the database or any of it's fields - I'm going off of an img. Let me know if this can be fixed and if populating dropdowns is setup. I'll be on later.